refactor: move the configuration of the theme
