+++ src/app/layout/layout.tsx
@@ -1,18 +1,12 @@
-import { FC, memo, useCallback, useContext } from 'react';
+import { FC, memo, useContext } from 'react';
 import { ILayoutProps } from '../types';
 import { Footer, Header } from '@/widgets';
 import { Outlet } from 'react-router-dom';
-import { ThemeContext } from '@/entities';
+import { ThemeContext } from '@/shared';
 import styles from './layout.module.css';
 
 export const Layout: FC<ILayoutProps> = memo(({ children }) => {
-  const { theme, setTheme } = useContext(ThemeContext);
-
-  const toggleTheme = useCallback(() => {
-    if (setTheme) {
-      setTheme(theme === 'light' ? 'dark' : 'light');
-    }
-  }, [setTheme, theme]);
+  const { theme } = useContext(ThemeContext);
 
   return (
     <div
@@ -20,7 +14,7 @@ export const Layout: FC<ILayoutProps> = memo(({ children }) => {
         ' '
       )}
     >
-      <Header setTheme={toggleTheme} theme={theme || ''} />
+      <Header />
       <div>{children}</div>
       <div>
         <Outlet />
