+++ src/entities/index.ts
@@ -1,4 +1,3 @@
-export { ThemeProvide, ThemeContext } from './theme';
 export { ErrorResponse } from './errorResponse';
 export { Card } from './card';
 export { Detail } from './detail';
