--- src/entities/theme/index.ts
@@ -1,2 +1,0 @@
-export { ThemeProvide } from './ui/themeProvide';
-export { ThemeContext } from './config/themeConfig';
