--- src/features/toggleTheme/model/toggleThemeTypes.ts
@@ -1,4 +1,0 @@
-export interface IToggleProps {
-  setTheme: () => void;
-  theme: string;
-}
