+++ src/shared/config/index.ts
@@ -1,0 +1,2 @@
+export { ThemeContext } from './theme/themeConfig';
+export type { IThemeContext } from './theme/themeConfig';
