+++ src/shared/index.ts
@@ -30,3 +30,5 @@ export type {
 } from './types';
 export { default as Sun } from './assets/icons/sun-svgrepo-com.svg?react';
 export { default as Moon } from './assets/icons/moon-svgrepo-com.svg?react';
+export { ThemeContext } from './config';
+export type { IThemeContext } from './config';
