--- src/widgets/header/model/headerTypes.ts
@@ -1,4 +1,0 @@
-export interface IHeaderProps {
-  setTheme: () => void;
-  theme: string;
-}
