+++ src/widgets/header/ui/header.tsx
@@ -1,14 +1,13 @@
 import { FC, memo } from 'react';
 import { Search, ToggleTheme } from '@/features';
 import styles from './header.module.css';
-import { IHeaderProps } from '../model/headerTypes';
 
-export const Header: FC<IHeaderProps> = memo(({ setTheme, theme }) => {
+export const Header: FC = memo(() => {
   return (
     <div className={styles.header}>
       <div className={styles.wrapper}>
         <Search />
-        <ToggleTheme setTheme={setTheme} theme={theme || ''} />
+        <ToggleTheme />
       </div>
     </div>
   );
